move statistics to their own file in utils
